Add `Visitor` Trait
Create the `syntax::Visitor` trait and implement it for the
evaluator. This moves the evaluator visiting logic out of the `eval`
module and into `syntax`
